requirements.txt
amqp==2.2.2
appnope==0.1.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
billiard==3.5.0.3
cachetools==2.1.0
cassandra-driver==3.14.0
celery==4.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cld2-cffi==0.1.4
cymem==1.31.2
Cython==0.28.2
cytoolz==0.9.0.1
dateparser==0.7.0
decorator==4.3.0
dill==0.2.7.1
Django==2.0.5
django-cors-headers==2.2.0
django-extensions==2.0.7
django-tastypie==0.14.1
dpath==1.4.2
elasticsearch==6.2.0
ftfy==4.4.3
GDAL==2.2.4
gunicorn==19.8.1
html5lib==1.0.1
idna==2.6
ijson==2.3
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
kombu==4.1.0
more-itertools==4.2.0
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
networkx==2.1
numpy==1.14.3
pandas==0.23.0
parso==0.2.0
pathlib==1.0.1
pexpect==4.5.0
pickleshare==0.7.4
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
pycparser==2.18
pyemd==0.5.1
Pygments==2.2.0
Pyphen==0.9.4
pytest==3.6.0
pytest-django==3.2.1
python-dateutil==2.7.3
python-Levenshtein==0.12.0
python-mimeparse==1.6.0
pytz==2018.4
redis==2.10.6
regex==2017.4.5
requests==2.18.4
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
spacy==2.0.11
termcolor==1.1.0
textacy==0.6.1
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.4
traitlets==4.3.2
tzlocal==1.5.1
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
